+++ src/App.tsx
@@ -2,8 +2,8 @@ import { Auth0Provider } from "@auth0/auth0-react";
 import "./App.css";
 import { Conditions } from "./components/content/conditions";
 import { PatientMedications } from "./components/content/medications/patient-medications";
-import { CTWProvider } from "./components/core/ctw-provider";
-import { PatientProvider } from "./components/core/patient-provider";
+import { CTWProvider } from "./components/core/providers/ctw-provider";
+import { PatientProvider } from "./components/core/providers/patient-provider";
 import { ErrorBoundary } from "./error-boundary";
 import { SecuredApp } from "./SecuredApp";
 import { PatientConditions } from "./components/content/conditions/patient-conditions";
