+++ src/api/autocomplete-conditions.ts
@@ -1,5 +1,5 @@
 import { getFormsConditionsUrl } from "./urls";
-import { CTWRequestContext } from "@/components/core/ctw-context";
+import { CTWRequestContext } from "@/components/core/providers/ctw-context";
 import { ctwFetch } from "@/utils/request";
 
 export const getAutoCompleteConditions = async (
