+++ src/components/content/condition-history/conditions-history.tsx
@@ -6,7 +6,7 @@ import {
   CollapsibleDataListStack,
   CollapsibleDataListStackEntries,
 } from "../../core/collapsible-data-list-stack";
-import { useCTW } from "../../core/ctw-provider";
+import { useCTW } from "../../core/providers/ctw-provider";
 import { CCDAModal } from "../../core/modal-ccda";
 import { RenderDocumentButton } from "../CCDA/render-document-button";
 import { ConditionHeader } from "../condition-header";
