+++ src/components/content/conditions/conditions.stories.tsx
@@ -8,8 +8,8 @@ import { emptyConditions } from "./story-helpers/mocks/empty-conditions";
 import { otherConditions } from "./story-helpers/mocks/other-conditions";
 import { patientConditions } from "./story-helpers/mocks/patient-conditions";
 import { setupConditionMocks } from "./story-helpers/mocks/requests";
-import { CTWProvider } from "@/components/core/ctw-provider";
-import { PatientProvider } from "@/components/core/patient-provider";
+import { CTWProvider } from "@/components/core/providers/ctw-provider";
+import { PatientProvider } from "@/components/core/providers/patient-provider";
 import { SYSTEM_ZUS_UNIVERSAL_ID } from "@/fhir/system-urls";
 
 type Props = ConditionsProps;
