+++ src/components/content/conditions/patient-conditions.stories.tsx
@@ -6,8 +6,8 @@ import {
 import { otherConditions } from "./story-helpers/mocks/other-conditions";
 import { patientConditions } from "./story-helpers/mocks/patient-conditions";
 import { setupConditionMocks } from "./story-helpers/mocks/requests";
-import { CTWProvider } from "@/components/core/ctw-provider";
-import { PatientProvider } from "@/components/core/patient-provider";
+import { CTWProvider } from "@/components/core/providers/ctw-provider";
+import { PatientProvider } from "@/components/core/providers/patient-provider";
 import { SYSTEM_ZUS_UNIVERSAL_ID } from "@/fhir/system-urls";
 
 type Props = PatientConditionsProps;
