+++ src/components/content/medications/medication-history.stories.tsx
@@ -7,8 +7,8 @@ import { otherProviderMedications } from "./story-helpers/mocks/other-provider-m
 import { providerMedications } from "./story-helpers/mocks/provider-medications";
 import { setupMedicationMocks } from "./story-helpers/mocks/requests";
 import { aggregatedFromMedStatement } from "@/components/content/medications/story-helpers/mocks/aggregated-from-med-statement";
-import { CTWProvider } from "@/components/core/ctw-provider";
-import { PatientProvider } from "@/components/core/patient-provider";
+import { CTWProvider } from "@/components/core/providers/ctw-provider";
+import { PatientProvider } from "@/components/core/providers/patient-provider";
 import { MedicationStatementModel } from "@/fhir/models";
 import { SYSTEM_ZUS_UNIVERSAL_ID } from "@/fhir/system-urls";
