+++ src/components/content/medications/other-provider-meds-table.stories.tsx
@@ -6,8 +6,8 @@ import {
 } from "@/components/content/medications/other-provider-meds-table";
 import { otherProviderMedications } from "@/components/content/medications/story-helpers/mocks/other-provider-medications";
 import { providerMedications } from "@/components/content/medications/story-helpers/mocks/provider-medications";
-import { CTWProvider } from "@/components/core/ctw-provider";
-import { PatientProvider } from "@/components/core/patient-provider";
+import { CTWProvider } from "@/components/core/providers/ctw-provider";
+import { PatientProvider } from "@/components/core/providers/patient-provider";
 import { SYSTEM_ZUS_UNIVERSAL_ID } from "@/fhir/system-urls";
 
 type Props = OtherProviderMedsTableProps;
