+++ src/fhir/client.ts
@@ -1,5 +1,5 @@
 import Client from "fhir-kit-client";
-import { Env } from "@/components/core/ctw-provider";
+import { Env } from "@/components/core/providers/ctw-provider";
 import { CTW_REQUEST_HEADER } from "@/utils/request";
 
 export function getFhirClient(
