+++ src/fhir/practitioner.ts
@@ -1,6 +1,6 @@
 import { Reference } from "fhir/r4";
 import { searchBuilderRecords } from "./search-helpers";
-import { CTWRequestContext } from "@/components/core/ctw-context";
+import { CTWRequestContext } from "@/components/core/providers/ctw-context";
 import { PractitionerModel } from "@/fhir/models/practitioner";
 import { claimsAuthEmail, claimsPractitionerId } from "@/utils/auth";
