+++ src/hooks/use-breakpoints.ts
@@ -2,7 +2,7 @@ import useResizeObserver from "@react-hook/resize-observer";
 import { mapValues } from "lodash";
 import { RefObject, useState } from "react";
 import { useIsomorphicLayoutEffect } from "./use-isomorphic-layout-effect";
-import { useCTW } from "@/components/core/ctw-provider";
+import { useCTW } from "@/components/core/providers/ctw-provider";
 import { defaultBreakpoints } from "@/styles/tailwind.theme";
 
 type BreakpointKeys = keyof typeof defaultBreakpoints;
