+++ src/index.ts
@@ -17,10 +17,10 @@ export * from "@/components/core/pagination/pagination";
 export * from "@/components/core/pagination/pagination-list";
 export * as CTWBox from "@/components/core/ctw-box";
 // Zus Providers and Hooks
-export * from "@/components/core/ctw-provider";
+export * from "@/components/core/providers/ctw-provider";
 export * from "@/components/core/drawer";
 export * from "@/components/core/modal";
-export * from "@/components/core/patient-provider";
+export * from "@/components/core/providers/patient-provider";
 export * from "@/components/core/spinner";
 export * from "@/components/core/table/table";
 export * from "@/components/core/toggle";
